Make the highlights container an island and begin adding buttons
